requirements.txt
Django==3.2.7
djangorestframework==3.12.4
python-decouple==3.5
psycopg2-binary
pillow
Django-filter
drf-yasg

